begin generalized modeling function to facilitate modeling across geographic levels (county and FRR) and predictor sets (Nolte, 2020 and Gold et al 2023).
The motivation for this function (laid out further in #51) is that the repetitive modeling scripts are highly inefficient and, more pressing, do not allow for quick, flexible adjustments to what is being modeled, using what data, and what statistics/results are saved where. For analysis discussed by Seth and Asa on 23 jan 2023 and referenced in #48, #49, and #50, we need to predict sales values of *all* observations, not just those in the testing set. Redesigning the modeling structure was not strictly necessary, but it will greatly improve efficiency later on and help avoid hard-to-track-down bugs.

The parent function (fmv_model()) dispatches to geo-specific methods for county and frr models, which both take arguments indicating which predictor set to use, as well as geo-specific arguments (HPI for county, only.nolte.counties for FRR).