This commit completes the main action described in #51 with the following:
1. Defining frr_model() as a method to model a single FRR. This will be mapped over by fmv_model.frr() to model all 9 FRRs. The basic structure is:
  a. Define model predictor set
  b. Read in data from all counties in current FRR using read_frr_data(), which ttakes an arg (only.nolte.counties) that filters down the returned data to the predefined set of counties modeled by Nolte's base county model.
  c. Filter data down to only include model predictors, outcome, and other relevant helper columns using frr_feature_selection()
  d. Model FRR using rf_fit() (frr method descibred more below)
  e. Parse results from model and return predictions (all), predicitions (just test), variable importance, and performance metrics).
  f. Save each result type in its appropriate folder.

The other major contribution of this commit is the move toward making "generic" functions that can be dispatched to either frr or county methods. For instance, rf_fit(), make_results_path(), and extract_results() now all have .frr and .county methods, or equivalent. This is done to simplify the function calls and (hopefully) make the frr_model() and model_state_counties() easier to read.