Major fix to how model predictions are saved in the new base_models paradigm, as requestd in #59.
Previously, test-set predictions for FRR models were bound to unique IDs (SIDs) using a custom function called bind_sid_to_pred(), which used to the .row column in the last_fit() .predictions output to match the predicted values with the original test-set dataframe.

This produced NA error in the SID variable, because the .row column from the last_fit() output was simply not correctly identifying the original test-set df rows (or identifying them in a manner I could not understand).

To simplify this prediction collection process and solve this error, I now call augment(last_fit) directly in extract_results.frr(), which attaches the predicted values to the original dataframe. We no longer get the .row variable, but this should not cause any issues downstream. I have tested this new approach on 500- and 1000-row samples in all 9 FRRs, and it has produced no NA values for the SID column.

Other fixes in this commit:
1. Makes back-compatible version of loadResults() to allow old archived results to be retrieved, using archive = TRUE. Also supports outmoded term "performance" (equivalent to new "metrics" language).
2. Remove MHV join in frr_feature_selection(), which was redundant, producing duplicated MHV variables with .x and .y suffixes